Godot and git clusterfuck
I checked out the main branch (while Godot was closed as the documentation recommends) and because the branch was very old, it lost almost all assets and scripts. This commit also is doing something to files but it says there are no changes? I don't know if I did something wrong or if this is a Godot bug.