Player UI started
Made a UI element in the bottom right corner to show the mag size and remaining shots in mag. Made a signal in the player script that emits every shot and that has the remaining shots as an argument. Also made an empty script in the player ui node that will update the numbers