Fixed last bug
Before, if you shot all the rounds in the mag it would automatically reload. Then every time you sprinted and cancelled the reload, it would never stop trying to reload after each sprint. Now when it is cancelled it doesn't try to reload again. It only reloads then if the player presses R or tries to fire