Moving up the ragdoll
Going from the feet up. In the character's script, I wrote a method that sets all of the physical bones visibility to true so that I can turn them off in the editor. This is so the gizmos don't clutter up the screen and I know which bones I have fixed