Commit before working on editor scripts
The gizmos are making it hard to work with the ragdoll. I have been toggling visibility of bones to make it easier but I can only do it one at a time and it is super annoying. Godot has ways of running scripts in the editor so I'm gonna try making a script for toggling node visibility faster