Started script and fixed armature
When exporting a gltf from blender, the last bone in a chain isn't in the armature. This is because the connection of two bones (joints) are the important part of the armature or something. So I added a leaf bone to the end of the armature in blender and reimported. Now there are the correct number of physical bones and I can start messing with the joints