I MIGHT BE FIXING IT
For some reason the basis of most of the joints were all screwed up? Thats what made the warning of non uniform scale pop up next to a lot of the joints. I am going through and fixing them all. Hopefully they stay that way and hopefully that was causing the spazzing